+++ polaris-react/src/components/Banner/Banner.tsx
@@ -26,7 +26,6 @@ import {Icon, IconProps} from '../Icon';
 import {WithinContentContext} from '../../utilities/within-content-context';
 import {Text} from '../Text';
 import {Box} from '../Box';
-import {Bleed} from '../Bleed';
 
 import styles from './Banner.scss';
 
@@ -174,10 +173,10 @@ export const Banner = forwardRef<BannerHandles, BannerProps>(function Banner(
           <Icon source={iconName} color={iconColor} />
         </Box>
 
-        <Bleed marginInline="0" marginBlockStart="05">
+        <div className={styles.ContentWrapper}>
           {headingMarkup}
           {contentMarkup}
-        </Bleed>
+        </div>
       </div>
     </BannerContext.Provider>
   );