+++ polaris.shopify.com/pages/examples/filters-with-help-text.tsx
@@ -2,7 +2,7 @@ import {
   ChoiceList,
   TextField,
   RangeSlider,
-  Card,
+  LegacyCard,
   ResourceList,
   Filters,
   Avatar,
@@ -134,7 +134,7 @@ function ResourceListFiltersExample() {
 
   return (
     <div style={{height: '568px'}}>
-      <Card>
+      <LegacyCard>
         <ResourceList
           resourceName={{singular: 'customer', plural: 'customers'}}
           filterControl={
@@ -182,7 +182,7 @@ function ResourceListFiltersExample() {
             );
           }}
         />
-      </Card>
+      </LegacyCard>
     </div>
   );