+++ polaris.shopify.com/pages/examples/index-table-with-bulk-actions.tsx
@@ -1,4 +1,4 @@
-import {IndexTable, Card, useIndexResourceState, Text} from '@shopify/polaris';
+import {IndexTable, LegacyCard, useIndexResourceState, Text} from '@shopify/polaris';
 import React from 'react';
 import {withPolarisExample} from '../../src/components/PolarisExampleWrapper';
 
@@ -71,7 +71,7 @@ function IndexTableWithBulkActionsExample() {
   );
 
   return (
-    <Card>
+    <LegacyCard>
       <IndexTable
         resourceName={resourceName}
         itemCount={customers.length}
@@ -90,7 +90,7 @@ function IndexTableWithBulkActionsExample() {
       >
         {rowMarkup}
       </IndexTable>
-    </Card>
+    </LegacyCard>
   );
 }