+++ polaris.shopify.com/pages/examples/index-table-with-row-status.tsx
@@ -1,4 +1,4 @@
-import {IndexTable, Card, useIndexResourceState, Text} from '@shopify/polaris';
+import {IndexTable, LegacyCard, useIndexResourceState, Text} from '@shopify/polaris';
 import React from 'react';
 import {withPolarisExample} from '../../src/components/PolarisExampleWrapper';
 
@@ -53,7 +53,7 @@ function IndexTableWithRowStatusExample() {
   );
 
   return (
-    <Card>
+    <LegacyCard>
       <IndexTable
         resourceName={resourceName}
         itemCount={customers.length}
@@ -70,7 +70,7 @@ function IndexTableWithRowStatusExample() {
       >
         {rowMarkup}
       </IndexTable>
-    </Card>
+    </LegacyCard>
   );
 }