+++ polaris.shopify.com/pages/examples/legacy-stack-spacing.tsx
@@ -1,0 +1,14 @@
+import {LegacyStack, Badge} from '@shopify/polaris';
+import React from 'react';
+import {withPolarisExample} from '../../src/components/PolarisExampleWrapper';
+
+function LegacyStackExample() {
+  return (
+    <LegacyStack spacing="loose">
+      <Badge>Paid</Badge>
+      <Badge>Fulfilled</Badge>
+    </LegacyStack>
+  );
+}
+
+export default withPolarisExample(LegacyStackExample);