+++ polaris.shopify.com/pages/examples/tabs-with-badge-content.tsx
@@ -1,4 +1,4 @@
-import {Badge, Card, Tabs} from '@shopify/polaris';
+import {Badge, LegacyCard, Tabs} from '@shopify/polaris';
 import {useState, useCallback} from 'react';
 import {withPolarisExample} from '../../src/components/PolarisExampleWrapper';
 
@@ -33,13 +33,13 @@ function TabsWithBadgeExample() {
   ];
 
   return (
-    <Card>
+    <LegacyCard>
       <Tabs tabs={tabs} selected={selected} onSelect={handleTabChange} fitted>
-        <Card.Section title={tabs[selected].content}>
+        <LegacyCard.Section title={tabs[selected].content}>
           <p>Tab {selected} selected</p>
-        </Card.Section>
+        </LegacyCard.Section>
       </Tabs>
-    </Card>
+    </LegacyCard>
   );
 }