+++ stylelint-polaris/index.js
@@ -4,6 +4,35 @@ const {
   tokens,
 } = require('@shopify/polaris-tokens');
 
+const disallowedUnits = [
+  'px',
+  'rem',
+  'em',
+  '%',
+  'ex',
+  'ch',
+  'lh',
+  'rlh',
+  'vw',
+  'vh',
+  'vmin',
+  'vmax',
+  'vb',
+  'vi',
+  'svw',
+  'svh',
+  'lvw',
+  'lvh',
+  'dvw',
+  'dvh',
+  'cm',
+  'mm',
+  'Q',
+  'in',
+  'pc',
+  'pt',
+];
+
 /**
  * @type {import('./plugins/coverage').PrimaryOptions} The stylelint-polaris/coverage rule expects a 3-dimensional rule config that groups Stylelint rules by coverage categories. It reports problems with dynamic rule names by appending the category to the coverage plugin's rule name
 
@@ -93,10 +122,11 @@ const stylelintPolarisCoverageOptions = {
       },
       'declaration-property-unit-disallowed-list': [
         {
-          '/^font/': ['px', 'rem', 'em'],
-          'line-height': ['px', 'rem', 'em'],
+          '/^font/': disallowedUnits,
+          'line-height': disallowedUnits,
         },
       ],
+      'property-disallowed-list': ['text-transform'],
       'function-disallowed-list': [
         'font-family',
         'font-size',
@@ -223,9 +253,9 @@ const stylelintPolarisCoverageOptions = {
       ),
       'declaration-property-unit-disallowed-list': [
         {
-          '/^padding/': ['px', 'rem', 'em'],
-          '/^margin/': ['px', 'rem', 'em'],
-          '/^gap/': ['px', 'rem', 'em'],
+          '/^padding/': disallowedUnits,
+          '/^margin/': disallowedUnits,
+          '/^gap/': disallowedUnits,
         },
       ],
       'polaris/global-disallowed-list': [
@@ -252,11 +282,11 @@ const stylelintPolarisCoverageOptions = {
       ].map(matchNameRegExp),
       'declaration-property-unit-disallowed-list': [
         {
-          'border-width': ['px', 'rem', 'em'],
-          border: ['px', 'rem', 'em'],
-          'border-radius': ['px', 'rem', 'em'],
-          'outline-offset': ['px', 'rem', 'em'],
-          outline: ['px', 'rem', 'em'],
+          'border-width': disallowedUnits,
+          border: disallowedUnits,
+          'border-radius': disallowedUnits,
+          'outline-offset': disallowedUnits,
+          outline: disallowedUnits,
         },
       ],
       'polaris/at-rule-disallowed-list': {
@@ -276,7 +306,7 @@ const stylelintPolarisCoverageOptions = {
         // Legacy custom properties
         // /--p-border-radius-base/,
         /--p-border-radius-wide/,
-        /--p-border-radius-full/,
+        // /--p-border-radius-full/,
         /--p-control-border-width/,
         /--p-thin-border-subdued/,
         /--p-banner-border-default/,
@@ -297,7 +327,7 @@ const stylelintPolarisCoverageOptions = {
       'function-disallowed-list': ['shadow'].map(matchNameRegExp),
       'declaration-property-unit-disallowed-list': [
         {
-          'box-shadow': ['px', 'rem', 'em'],
+          'box-shadow': disallowedUnits,
         },
       ],
       'property-disallowed-list': ['text-shadow'],